harden firmware update validation and display
Improve the standalone update page, hide online updates in AP mode, and report upload progress with a persistent LCD status screen. Use the ESP8266 stack thunk for catalog signature verification to avoid exhausting the normal stack. Require manual uploads to provide an exact size and browser-computed SHA-256, verify the digest while streaming, and finalize the image only after all checks pass.

+++ OpenSprinkler.h
@@ -459,7 +459,7 @@ public:
 	static void lcd_print_time(time_os_t t);  // print current time
 	static void lcd_print_ip(const unsigned char *ip, unsigned char endian);  // print ip
 	static void lcd_print_mac(const unsigned char *mac);  // print mac
-	static void lcd_print_update(const char *message, uint8_t percent); // print firmware update progress
+	static void lcd_print_update(const char *message, int16_t percent); // percent < 0 means indeterminate
 	static void lcd_print_screen(char c);  // print station bits of the board selected by display_board
 	static void lcd_print_version(unsigned char v);  // print version number
 	static void lcd_set_brightness(unsigned char value=1);
